requirements.txt
aiohttp==3.7.3
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==3.0.1
asyncio==3.4.3
asyncpg==0.21.0
attrs==20.3.0
backcall==0.2.0
bleach==3.2.1
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
git+https://github.com/Rapptz/discord-ext-menus
discord.py==1.6.0
docutils==0.16
entrypoints==0.3
jsonschema==3.2.0
typing-extensions==3.7.4.3